Mudificações em Printer e Cronometro

+++ CalculadoraUniversal/src/Datas/CalculoLDT.java
@@ -19,6 +19,7 @@ import java.time.temporal.ChronoUnit;
 import java.time.temporal.TemporalAccessor;
 import java.util.Optional;
 import java.util.OptionalInt;
+import java.util.OptionalLong;
 
 /**
  *
@@ -168,7 +169,7 @@ public class CalculoLDT implements ICalculoLDT, ICalculoTempo {
 	}
 
 
-	public OptionalInt semanasDesdeInicio() {
+	public OptionalLong semanasDesdeInicio() {
 		return datas.semanasDesdeInicio();
 	}
